+++ CHANGELOG.md
@@ -2,11 +2,10 @@
 
 ## v.1.3.3
 
-- add initial support for moonsault create GUI
 - add moonsault.currentRoute and moonsault.previousRoute properties
 - add moonsault.urlParams.set and moonsault.urlParams.delete for managing URL parameters
 - add moonsault.buildURL for building paramater list and and navigating to a route
-- add support for data-hash-change property on components
+- add support for data-hash-change property on components.
 
 ## v.1.3.2